requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
wikipedia
googletrans==4.0.0-rc1
gtts
pygame
google-generativeai==0.5.2
